[HomePage] Added HomeNewsPanel and some mock news cards - Part 2
