+++ app/components/HomeNewsPanel/News/Info/index.js
@@ -11,10 +11,8 @@ import Text from 'material-ui/Text';
 function Changelog2() {
   return (
     <Text component="p">
-      + Dodany news panel na stronie głównej <br />
-      + Dodany prototyp searchera do app baru <br />
-      + Dodany prototyp zegaru systemowego <br />
-      + Bug Fix <br />
+      Od 20.03.2017r. do nieokreślonego dnia w przyszłości system będzie niedostępny. Projekt jest wciąż w fazie
+      pre-alfa prototypu i może nie być gotowy do użytku. Prosimy o zrozumienie.
     </Text>
   );
 }
