+++ app/containers/HomePage/index.js
@@ -15,6 +15,9 @@ import Layout from 'material-ui/Layout';
 import StatusPanel from '../../components/StatusPanel';
 import InfoPanel from '../../components/InfoPanel';
 import HomeNewsPanel from '../../components/HomeNewsPanel';
+import Changelog1 from '../../components/HomeNewsPanel/News/Changelog1';
+import Changelog2 from '../../components/HomeNewsPanel/News/Changelog2';
+import Info from '../../components/HomeNewsPanel/News/Info';
 
 function HomePage() {
   return (
@@ -29,7 +32,13 @@ function HomePage() {
           <StatusPanel />
         </Layout>
         <Layout item xs={12}>
-          <HomeNewsPanel />
+          <HomeNewsPanel date="20.03.2017" title="Informacje Bieżące" content={<Info />} />
+        </Layout>
+        <Layout item xs={12}>
+          <HomeNewsPanel date="20.03.2017" title="Changelog #2" content={<Changelog2 />} />
+        </Layout>
+        <Layout item xs={12}>
+          <HomeNewsPanel date="15.03.2017" title="Changelog #1" content={<Changelog1 />} />
         </Layout>
       </Layout>
     </Paper>
